requirements.txt
Django>=5.0,<6.0
python-dotenv>=1.0
requests>=2.31
beautifulsoup4>=4.12
psycopg[binary]>=3.2 ; platform_system != 'Windows'
django-environ>=0.11
pytest>=8.0
pytest-django>=4.8